requirements.txt
Flask
numpy
pandas
scikit_learn
tensorflow-cpu
gunicorn
pickle-mixin